controller/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
typing_extensions
httpx
kubernetes
pyyaml
